Generalizacion. Extraemos tests
