+++ complejidad-computacional-metodos-de-ordenamiento/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>complejidad-computacional-metodos-de-ordenamiento-elementales</name>
+	<name>complejidad-computacional-metodos-de-ordenamiento</name>
 	<comment></comment>
 	<projects>
 	</projects>
